ml_engineer2/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
psycopg2-binary
pgvector
numpy
python-dotenv
google-generativeai
PyMuPDF
pdfplumber
python-docx
extract-msg
pdf2image
pytesseract
langchain
langchain-text-splitters
